Replace bands column revision with email_id This revision is usale out of the box, unlike the last one, which we also don't need anymore.
