+++ alembic/versions/6a885e4ac48d_add_email_id_to_email_table.py
@@ -1,0 +1,36 @@
+"""Add email_id to Email table
+
+Revision ID: 6a885e4ac48d
+Revises: 
+Create Date: 2016-06-02 23:18:41.979751
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = '6a885e4ac48d'
+down_revision = None
+branch_labels = None
+depends_on = None
+
+from alembic import op
+import sqlalchemy as sa
+import sideboard
+
+
+def upgrade():
+    ### commands auto generated by Alembic - please adjust! ###
+    email = sa.table('email', sa.Column('email_id', sa.String), sa.Column('subject', sa.String))
+
+    op.add_column('email', sa.Column('email_id', sideboard.lib.sa.CoerceUTF8(), server_default='', nullable=False))
+    op.execute(
+        email.update().where(email.c.email_id=='').values(
+            email_id=email.c.subject
+        )
+    )
+    ### end Alembic commands ###
+
+
+def downgrade():
+    ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('email', 'email_id')
+    ### end Alembic commands ###
